+++ CHANGELOG.md
@@ -10,6 +10,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2476 Use tide-whisperer v2: route /all.
 - YLP-2696 Fix some Sonarcloud bugs and code smells
 - YLP-2704 Remove settings logic from viz
+- YLP-2709 migration warm up tooltip
 
 ## 3.0.5-beta - 2023-07-12
 ### Added